fix multidrag action
